simple_form: Add paging to main page + debug(True)
This adds a couple of dynamic paging controls to the HTML around the
Send new notification link (<< and >>).  It also adds a little note
at the bottom to indicate the number of pages, if that exceeds one page.

These controls are greyed out at the extrema of the pages. So they're
both rendered as plain <H3> text when there's only one page full of
open entries, the << is only rendered as a link when viewing pages
greater than zero and the >> os only rendered as a link when there
are more pages available to page to.

In the Model.get_open_entries() method I return a pagemax value as
well as the number of rows requested (20 by default).  This entailed
adding the get_open_entry_count() method and thus allows the root
route to render these additional HTML page elements.

Also enabled debug(True) properly.