Model: message threads for display and close
In the Model() the the get_open_entries() method
fetches the tail of each thread, and get_open_entry_count()
return the number of threads rather than the number
of message entries.  The new get_thread_entries() method
gets a list of IDs for all messages in a thread, given *any*
message therein.  (It walks up the tree to the root and then
recurses downward through all branches).  close_entry() renamed
to close_thread() and, as the name suggests, now closes an
entire thread

This all relies on the schema changes in the previous
commit.  No further schema changes were made.