Adding login/cooking support
I've added simplistic (and not fully functional) support for a
settings.ini file and modular authentication support driven by
that configuration.

Along with this I've added mockauth.py as such a module (and
with the ability to create accounts for testing purposes).

Currently an authentication module must:
 * provide an Auth class or function
 * accept the name of a Config.ini file to read for options
 * return an object which implements an authenticate method
 * authenticate must take name, (purported) password as arguments
 * return the name if authenticated or else '' (empty string)

I think I actually to change that to return name and session ID

For now the cookie being set is all wrong.  It needs to be a
session ID so I can get the username from the session DB.
Also I need to set with expiration and probably use a nonce to
"armor" it.

I'll read the tutorial on using "beaker" session management with
Bottle before I continue.